+++ features/paper/hooks/use-papers.ts
@@ -1,25 +1,23 @@
 import { useCallback, useState } from "react";
-
 import { Paper } from "@/features/paper/types";
 
-export default function useNodePapers() {
-  const [paperIndexing, setPaperIndexing] = useState<
+export default function usePapers(initialPapers?: Paper.Scheme.Metadata[]) {
+  const [cache, setCache] = useState<
     Map<Paper.Scheme.Id, Paper.Scheme.Metadata>
-  >(new Map());
+  >(new Map(initialPapers?.map((value) => [value.id, value])));
 
   const hasPaper = useCallback(
-    (paperID: Paper.Scheme.Id) => paperIndexing.has(paperID),
-    [paperIndexing],
+    (paperID: Paper.Scheme.Id) => cache.has(paperID),
+    [cache],
   );
 
   const getPaper = useCallback(
-    (paperID?: Paper.Scheme.Id) =>
-      paperID ? paperIndexing.get(paperID) : undefined,
-    [paperIndexing],
+    (paperID?: Paper.Scheme.Id) => (paperID ? cache.get(paperID) : undefined),
+    [cache],
   );
 
   const upsertPaper = useCallback((paper: Paper.Scheme.Metadata) => {
-    setPaperIndexing((papers) => {
+    setCache((papers) => {
       const newPapers = new Map(papers);
       newPapers.set(paper.id, paper);
 
@@ -31,7 +29,7 @@ export default function useNodePapers() {
     (paperID: Paper.Scheme.Id) => {
       if (!hasPaper(paperID)) return false;
 
-      setPaperIndexing((papers) => {
+      setCache((papers) => {
         const newPapers = new Map(papers);
         newPapers.delete(paperID);
